+++ libnavui-androidauto/src/main/java/com/mapbox/androidauto/car/routes/RoutesListener.kt
@@ -1,7 +1,0 @@
-package com.mapbox.androidauto.car.routes
-
-import com.mapbox.navigation.base.route.NavigationRoute
-
-internal fun interface RoutesListener {
-    fun onRoutesChanged(routes: List<NavigationRoute>)
-}