+++ libnavui-androidauto/src/main/java/com/mapbox/androidauto/car/routes/RoutesProvider.kt
@@ -1,6 +1,0 @@
-package com.mapbox.androidauto.car.routes
-
-internal interface RoutesProvider {
-    fun registerRoutesListener(listener: RoutesListener)
-    fun unregisterRoutesListener(listener: RoutesListener)
-}